allow text/markdown for script type attribute
The parser needs a way to identify which scripts need to be processed. The best way to do this is with the type attribute. The text/markdown value is already in widespread use and should be supported, but we'll also support the older idiosyncratic literate-javascript for cases where it might be helpful to distinguish between Markdown that is being loaded as a document versus Markdown that is intended as literate code for execution.